Fix windows build.

+++ Engine/src/engine/render/Material.cpp
@@ -35,8 +35,8 @@ namespace se::render
         color.resize(lightSetup.pointLights.size());
         std::transform(lightSetup.pointLights.begin(), lightSetup.pointLights.end(), color.begin(), [](const PointLight& light){ return light.color; });
 
-        SetUniform("lightPos", asset::shader::ast::AstType::Vec3, pos.size(), &pos[0]);
-        SetUniform("lightColor", asset::shader::ast::AstType::Vec3, color.size(), &color[0]);
+        SetUniform("lightPos", asset::shader::ast::AstType::Vec3, static_cast<int>(pos.size()), &pos[0]);
+        SetUniform("lightColor", asset::shader::ast::AstType::Vec3, static_cast<int>(color.size()), &color[0]);
     }
 
     void Material::CreatePlatformResources(const VertexBuffer&)
